Extract discoverHomesets to ServiceRefresher
